bitdi

+++ Foxoft/UcReturn.cs
@@ -124,7 +124,7 @@ namespace Foxoft
 
             if (sumNetAmount != 0)
             {
-                int paymentType = 0;
+                byte paymentType = 0;
 
                 SimpleButton simpleButton = sender as SimpleButton;
                 switch (simpleButton.Name)
